+++ apps/frontend/src/devices/detail.tsx
@@ -4,7 +4,7 @@ import { LayoutHeader } from '../components/LayoutHeader';
 import { Column } from '../components/Column';
 import { Card } from '../components/Card';
 import { ErrorState } from '../components/AsyncUtils';
-import { Deployment, Device, Image, ImageVersion } from '../../types';
+import { Deployment, Device, CreateDeploymentDto } from '../types';
 import { DateTime } from '../components/DateTime';
 import { Pill } from '../components/Pill';
 import { toSentenceCase } from '../utils';
@@ -14,12 +14,7 @@ import { DetailField } from '../components/DetailView';
 import { CreateForm, Field } from '../components/CreateForm';
 import { CreateDialog } from '../components/CreateDialog';
 import { UseFormRegister, UseFormWatch } from 'react-hook-form';
-import { useQuery } from '@tanstack/react-query';
-import axios from 'axios';
-import { BASE_API_URL } from '../consts';
-import { LoadingState } from '../components/AsyncUtils';
-import { EmptyState } from '../components/AsyncUtils';
-import { Select } from '../components/FormUtils';
+import { ImageSelect, VersionSelect } from '../components/FormUtils';
 
 const deviceDetailFields: DetailField<Device>[] = [
   {
@@ -47,62 +42,6 @@ const deviceDetailFields: DetailField<Device>[] = [
     accessor: (device: Device) => <Pill>{toSentenceCase(device.state)}</Pill>
   }
 ];
-
-interface CreateDeploymentDto {
-  deviceId: string;
-  imageId: string;
-  imageVersionId: string;
-}
-
-
-
-const ImageSelect = ({ register }: { register: UseFormRegister<CreateDeploymentDto> }) => {
-  const { data: images, isLoading, isError } = useQuery({
-    queryKey: ['images'],
-    queryFn: () => axios.get(`${BASE_API_URL}/images`).then(res => res.data)
-  });
-
-  if (isLoading) return <LoadingState />;
-  if (isError) return <ErrorState>Error loading images</ErrorState>;
-  if (!images?.length) return <EmptyState>No images found</EmptyState>;
-
-  return (
-    <Select {...register('imageId')}>
-      <option value=''>Select an image</option>
-      {images.map((image: Image) => (
-        <option key={image.uuid} value={image.uuid}>
-          {image.id}
-        </option>
-      ))}
-    </Select>
-  );
-};
-
-const VersionSelect = ({ register, watch }: { register: UseFormRegister<CreateDeploymentDto>, watch?: UseFormWatch<CreateDeploymentDto> }) => {
-  const selectedImageId = watch ? watch('imageId') : '';
-
-  const { data: versions, isLoading, isError } = useQuery({
-    queryKey: ['image', selectedImageId, 'versions'],
-    queryFn: () => axios.get(`${BASE_API_URL}/images/${selectedImageId}/versions`).then(res => res.data),
-    enabled: !!selectedImageId
-  });
-
-  if (!selectedImageId) return <EmptyState>Select an image first</EmptyState>;
-  if (isLoading) return <LoadingState />;
-  if (isError) return <ErrorState>Error loading versions</ErrorState>;
-  if (!versions?.length) return <EmptyState>No versions available</EmptyState>;
-
-  return (
-    <Select {...register('imageVersionId')}>
-      <option value=''>Select a version</option>
-      {versions.map((version: ImageVersion) => (
-        <option key={version.uuid} value={version.uuid}>
-          {version.id}
-        </option>
-      ))}
-    </Select>
-  );
-};
 
 const deploymentFields: Field<CreateDeploymentDto>[] = [
   {
